requirements.txt

streamlit
pandas
pillow
xlsxwriter